requirements.txt
uv
dotenv
fastapi[standard]
mcp
langchain
langchain_mcp_adapters
langgraph
langchain_google_genai

pymongo

passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography] 
email-validator

pandas
pymysql
openpyxl

google.generativeai as genai
PyPDF2